+++ Magento/Mtf/Client/Driver/Selenium/Driver.php
@@ -417,6 +417,17 @@ class Driver implements DriverInterface
     }
 
     /**
+     * Hover mouse over an element.
+     *
+     * @param ElementInterface $element
+     * @return void
+     */
+    public function hover(ElementInterface $element)
+    {
+        $this->driver->moveto($this->getNativeElement($element));
+    }
+
+    /**
      * Send a sequence of key strokes to the active element.
      *
      * @param ElementInterface $element